LayoutTests/http/wpt/wk-web-animations/timing-model/animation-playback-rate.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Tests for playback rate</title>
<link rel="help" href="https://w3c.github.io/web-animations/#updating-the-playback-rate-of-an-animation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
  const animation = new Animation;

  assert_equals(animation.playbackRate, 1);
}, 'Playback rate defaults to 1');

test(t => {
  const animation = new Animation;
  animation.startTime = 1000;
  animation.playbackRate = 2;
  internals.setTimelineCurrentTime(document.timeline, 3000);

  assert_equals(animation.startTime, 1000);
  assert_equals(animation.currentTime, 4000);
  assert_equals(animation.playbackRate, 2);
}, 'Correctly compute the current time based on the playback rate');

test(t => {
  const animation = new Animation;
  animation.startTime = 2000;
  animation.playbackRate = 2;
  internals.setTimelineCurrentTime(document.timeline, 4000);

  animation.currentTime = 3000;

  assert_equals(animation.startTime, 2500);
  assert_equals(animation.currentTime, 3000);
  assert_equals(animation.playbackRate, 2);
}, 'Correctly compute the start time based on the playback rate when setting current time');

test(t => {
  const animation = new Animation;
  animation.startTime = 2000;
  animation.playbackRate = 0.5;
  internals.setTimelineCurrentTime(document.timeline, 4000);

  animation.currentTime = 1000;

  assert_equals(animation.startTime, 2000);
  assert_equals(animation.currentTime, 1000);
  assert_equals(animation.playbackRate, 0.5);
}, 'Correctly compute the start time based on a playback rate < 1 when setting current time');

test(t => {
  const animation = new Animation;
  animation.currentTime = 1000;
  animation.playbackRate = -1;
  internals.setTimelineCurrentTime(document.timeline, 4000);

  assert_equals(animation.startTime, 5000);
  assert_equals(animation.currentTime, 1000);
  assert_equals(animation.playbackRate, -1);
}, 'Correctly compute the start time based on a playback rate < 0');

test(t => {
  const animation = new Animation;
  animation.playbackRate = 0;
  animation.startTime = 0;
  internals.setTimelineCurrentTime(document.timeline, 1000);

  assert_equals(animation.startTime, 0);
  assert_equals(animation.currentTime, 0);
  assert_equals(animation.playbackRate, 0);
}, 'Correctly compute a current time of 0 when playback rate is 0');

</script>
</body>
